requirements.txt
Flask==3.0.3
gunicorn==22.0.0
Jinja2==3.1.3
Werkzeug==3.0.3
itsdangerous==2.2.0
click==8.1.7
requests==2.31.0
numpy==1.26.4
python-dotenv
flask_session
